fix(import): return 409 for invalid state transitions on start/complete
/start and /complete always returned 404 for any ValueError, including
state-transition conflicts (e.g. RUNNING -> RUNNING). Extract the
not-found-vs-conflict branching already used by /execute into a shared
helper and apply it to all three endpoints.

Also fix ImportJobResponse.progress: it was typed int while the model
column is Float, which would fail validation on any fractional value.

+++ backend/app/schemas/import_job.py
@@ -15,7 +15,7 @@ class ImportJobResponse(BaseModel):
     source_path: str
     source_type: str
     status: str
-    progress: int
+    progress: float
     files_discovered: int
     files_processed: int
     error_message: str | None
